aws/request: Overwrite (web http)'s DEFAULT-VAL-WRITER.
The DEFAULT-VAL-WRITER procedure in Guile's (web http) module causes
values in headers to be quoted when they contain a semicolon.

The Authorization header that AWS expects contains a field
SignedHeaders whose value is a semicolon-separated list of header
names.  When this list is quoted AWS considers the opening quote
character to be part of the name of the first header and the closing
quote character to be part of the name of the last header.

* aws/request.scm (my-default-val-writer): New procedure; replace the
definition of DEFAULT-VAL-WRITER in Guile's (web http).