base: aws-operation: Pass request arguments to request constructor.
Instead of expecting a single input of the required type, the
procedure returned by AWS-OPERATION now accepts keyword arguments
intended for the request constructor.  This makes for a much less
verbose DSL.  Compare the explicit style

  (DeleteFileSystem
    (DeleteFileSystemRequest
      #:FileSystemId "fs-cabba9e"))

with the new implicit style:

  (DeleteFileSystem
   #:FileSystemId "fs-cabba9e")

* aws/base.scm (aws-operation): Accept an input-constructor; apply it
to a provided list of keyword arguments.
* language/aws/spec.scm (compile-operation): Generate code that
specifies the input constructor.
* README.org: Update examples.